break SNS tests into appender-only and writer-only
