+++ appenders/src/test/resources/TestSNSAppender/testStaticClientFactory.properties
@@ -1,4 +1,4 @@
-# configuration for mock-writer test
+# configuration for client factory test
 
 log4j.rootLogger=DEBUG, default
 
@@ -8,4 +8,4 @@ log4j.appender.default.layout.ConversionPattern=%m
 
 log4j.appender.default.topicName=example
 
-log4j.appender.default.clientFactory=com.kdgregory.log4j.aws.TestSNSAppender.createMockClient
+log4j.appender.default.clientFactory=com.kdgregory.log4j.aws.TestSNSLogWriter.createMockClient
